refactor(mastracode): group platform integrations
Keep platform-backed GitHub and Linear implementations under a shared top-level platform directory so they sit alongside, rather than inside, the direct provider integrations.
