+++ mastracode/web/src/web/platform/github/event-worker.lifecycle.test.ts
@@ -4,9 +4,9 @@ import { LibSQLFactoryStorage } from '@mastra/libsql';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { MastraFactory } from '../../factory-entry.js';
+import { subscribeToPullRequest } from '../../github/subscriptions.js';
 import { __resetRuntimeConfigForTests } from '../../runtime-config.js';
 import { PlatformGithubIntegration } from './integration.js';
-import { subscribeToPullRequest } from '../subscriptions.js';
 
 const harness = vi.hoisted(() => {
   let mastra: Mastra | undefined;
